Revise talk details for Hidde van der Ploeg
Updated the speaker name and talk title, and revised the summary and content to reflect a focus on AI in app development.

Update speaker image for 2025 Day 2 Talk 6

Replaced the image for the 'Less AI, more Magic' talk from pol.webp to hidde.webp and updated the agenda entry accordingly.

Add Hidde as a speaker.

Rename file.